Implemented Meal entity

+++ src/main/java/com/project/foodinventory/repositories/FoodRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.foodinventory.models.Food;
 
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
 
 }
